pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whatsapp-analyzer"
version = "0.1.0"
description = "WhatsApp Group Chat Analyzer Agent - local LLM powered analysis"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "ollama>=0.4",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "pytesseract>=0.3",
    "Pillow>=10.0",
    "streamlit>=1.38",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
wa-analyze = "whatsapp_analyzer.interfaces.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
